Use LF consistently
Currently, the extensions mix CRLF and LF line endings. This obviously interferes with checksums and could cause other issues elsewhere. As such, this MR converts some files with CRLF line endings to LF and adds a `.gitattributes` file to ensure the correct line endings are used in the future.

This was extracted from !526, which seems to be stuck in review limbo for now. The changes to lua files only change line endings, not the content. The GitLab diff view might not show that properly, but IDEA should.